+++ src/core/config/AppearanceConfig.cpp
@@ -4,10 +4,13 @@
 #include "core/config/keys.hpp"
 
 #include <QMetaEnum>
+#include <QtDebug>
+#include <QFontDatabase>
 
 AppearanceConfig::AppearanceConfig() :
-    mPalette(),
-    mFont(),
+    mPatternGridFont(),
+    mPatternGridHeaderFont(),
+    mOrderGridFont(),
     mShowFlats(false),
     mShowPreviews(true)
 {
@@ -21,22 +24,54 @@ bool AppearanceConfig::showPreviews() const {
     return mShowPreviews;
 }
 
-QFont AppearanceConfig::font() const {
-    return mFont;
+QFont AppearanceConfig::patternGridFont() const {
+    return mPatternGridFont;
 }
 
-QFont& AppearanceConfig::font() {
-    return mFont;
+
+QFont AppearanceConfig::orderGridFont() const {
+    return mOrderGridFont;
 }
 
+QFont AppearanceConfig::patternGridHeaderFont() const {
+    return mPatternGridHeaderFont;
+}
+
+void AppearanceConfig::setPatternGridFont(const QFont &font) {
+    mPatternGridFont = font;
+}
+
+void AppearanceConfig::setPatternGridHeaderFont(const QFont &font) {
+    mPatternGridHeaderFont = font;
+}
+
+void AppearanceConfig::setOrderGridFont(const QFont &font) {
+    mOrderGridFont = font;
+}
+
+void AppearanceConfig::setShowFlats(bool showFlats) {
+    mShowFlats = showFlats;
+}
+
+void AppearanceConfig::setShowPreviews(bool showPreviews) {
+    mShowPreviews = showPreviews;
+}
 
 void AppearanceConfig::readSettings(QSettings &settings) {
 
     settings.beginGroup(Keys::Appearance);
 
-    auto fontName = settings.value(Keys::font, QStringLiteral("Cascadia Mono")).toString();
-    mFont.setFamily(fontName);
-    mFont.setPointSize(settings.value(Keys::fontSize, 12).toInt());
+    auto setFont = [](QFont &font, QString const& fontstr) {
+        if (fontstr.isEmpty() || !font.fromString(fontstr)) {
+            // use built-in as a default
+            font.setFamily("CascadiaMono");
+            font.setPointSize(12);
+        }
+    };
+
+    setFont(mPatternGridFont, settings.value(Keys::patternGridFont).toString());
+    setFont(mOrderGridFont, settings.value(Keys::orderGridFont).toString());
+    setFont(mPatternGridHeaderFont, settings.value(Keys::patternGridHeaderFont).toString());
 
     mShowFlats = settings.value(Keys::showFlats, mShowFlats).toBool();
     mShowPreviews = settings.value(Keys::showPreviews, mShowPreviews).toBool();
@@ -49,8 +84,9 @@ void AppearanceConfig::writeSettings(QSettings &settings) const {
 
     settings.beginGroup(Keys::Appearance);
 
-    settings.setValue(Keys::font, mFont.family());
-    settings.setValue(Keys::fontSize, mFont.pointSize());
+    settings.setValue(Keys::patternGridFont, mPatternGridFont.toString());
+    settings.setValue(Keys::orderGridFont, mOrderGridFont.toString());
+    settings.setValue(Keys::patternGridHeaderFont, mPatternGridHeaderFont.toString());
     settings.setValue(Keys::showFlats, mShowFlats);
     settings.setValue(Keys::showPreviews, mShowPreviews);