+++ src/forms/ConfigDialog.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "core/Config.hpp"
+#include "widgets/config/AppearanceConfigTab.hpp"
 #include "widgets/config/MidiConfigTab.hpp"
 #include "widgets/config/SoundConfigTab.hpp"
 
@@ -56,6 +57,7 @@ private:
 
     QVBoxLayout mLayout;
         QTabWidget mTabs;
+            AppearanceConfigTab mTabAppearance;
             MidiConfigTab mTabMidi;
             SoundConfigTab mTabSound;
         QDialogButtonBox mButtons;